split configure helpers and prepare helpers
