+++ core/modules/catalog/helpers/prepare/index.ts
@@ -1,67 +1,13 @@
-import { hasConfigurableChildren } from '../';
-
-/**
- * set product quantity to 1
- */
-function setDefaultQty (product) {
-  if (!product.qty) {
-    product.qty = 1
-  }
-}
-
-/**
- * set parent sku, this is needed, because in configuration process we will override sku by configurable_children.sku
- */
-function setParentSku (product) {
-  if (!product.parentSku) {
-    product.parentSku = product.sku
-  }
-}
-
-/**
- * Default object that are used in vsf
- */
-function setDefaultObjects (product) {
-  product.errors = {}; // this is an object to store validation result for custom options and others
-  product.info = {};
-}
-
-/**
- * Fill custom attributes for every configurable child
- */
-function setCustomAttributesForChild (product) {
-  if (!hasConfigurableChildren(product)) return
-  // handle custom_attributes for easier comparing in the future
-  product.configurable_children.forEach((child) => {
-    let customAttributesAsObject = {}
-    if (child.custom_attributes) {
-      child.custom_attributes.forEach((attr) => {
-        customAttributesAsObject[attr.attribute_code] = attr.value
-      })
-      // add values from custom_attributes in a different form
-      Object.assign(child, customAttributesAsObject)
-    }
-  })
-}
-
-/**
- * Init product_option, needed to next configuration step
- */
-function setDefaultProductOptions (product) {
-  if (product.product_option) return
-  product.product_option = {
-    extension_attributes: {
-      custom_options: [],
-      configurable_item_options: [],
-      bundle_options: []
-    }
-  }
-}
+import setDefaultQty from './setDefaultQty';
+import setDefaultObjects from './setDefaultObjects';
+import setParentSku from './setParentSku';
+import setCustomAttributesForChild from './setCustomAttributesForChild';
+import setDefaultProductOptions from './setDefaultProductOptions';
 
 /**
  * Take products and apply base modification to it
  */
-export function preConfigureProduct (product) {
+function preConfigureProduct (product) {
   // base product modifications
   setDefaultQty(product)
   setDefaultObjects(product)
@@ -81,4 +27,7 @@ function prepareProducts (products) {
   return preparedProducts
 }
 
-export default prepareProducts
+export {
+  prepareProducts,
+  preConfigureProduct
+}
