+++ core/modules/catalog/helpers/prepare/setDefaultObjects.ts
@@ -1,0 +1,7 @@
+/**
+ * Default object that are used in vsf
+ */
+export default function setDefaultObjects (product) {
+  product.errors = {}; // this is an object to store validation result for custom options and others
+  product.info = {};
+}
