+++ core/modules/catalog/helpers/prepare/setDefaultQty.ts
@@ -1,0 +1,8 @@
+/**
+ * set product quantity to 1
+ */
+export default function setDefaultQty (product) {
+  if (!product.qty) {
+    product.qty = 1
+  }
+}
